# app/Http/Controllers/Api/WorkOrderController.php
<?php

declare(strict_types=1);

namespace App\Http\Controllers\Api;

use App\Exceptions\InsufficientStockException;
use App\Exceptions\InvalidStateException;
use App\Http\Controllers\Controller;
use App\Http\Requests\Api\WorkOrder\StoreWorkOrderRequest;
use App\Models\Inventory\Product;
use App\Models\Inventory\StockAdjustment;
use App\Models\Inventory\WorkOrder;
use App\Models\Inventory\WorkOrderItem;
use App\Services\ProductLocationStockService;
use App\Services\WorkOrderService;
use App\Support\SequenceNumberRetry;
use Dedoc\Scramble\Attributes\QueryParameter;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\DB;

/**
 * @tags Work Orders
 */
class WorkOrderController extends Controller
{
    /**
     * List work orders.
     */
    #[QueryParameter('search', description: 'Search by WO number, product name, or SKU', type: 'string')]
    #[QueryParameter('status', description: 'Filter by status', type: 'string', enum: ['draft', 'pending', 'in_progress', 'completed', 'cancelled'])]
    #[QueryParameter('sort_by', description: 'Sort field (default: created_at)', type: 'string')]
    #[QueryParameter('sort_dir', description: 'Sort direction: asc or desc (default: desc)', type: 'string', enum: ['asc', 'desc'])]
    #[QueryParameter('per_page', description: 'Items per page (default: 15, max: 100)', type: 'integer')]
    public function index(Request $request): JsonResponse
    {
        $organizationId = $request->user()->organization_id;

        $query = WorkOrder::with(['product:id,name,sku,thumbnail', 'creator:id,name'])
            ->forOrganization($organizationId)
            ->when($request->input('search'), function ($query, $search) {
                $query->where(function ($q) use ($search) {
                    $q->where('work_order_number', 'like', "%{$search}%")
                        ->orWhereHas('product', function ($q2) use ($search) {
                            $q2->where('name', 'like', "%{$search}%")
                                ->orWhere('sku', 'like', "%{$search}%");
                        });
                });
            })
            ->when($request->input('status'), function ($query, $status) {
                $query->byStatus($status);
            });

        // Sorting (allowlist to prevent SQL injection)
        $allowedSortColumns = ['created_at', 'updated_at', 'work_order_number', 'status', 'quantity'];
        $sortBy = in_array($request->input('sort_by'), $allowedSortColumns) ? $request->input('sort_by') : 'created_at';
        $sortDir = ($request->input('sort_dir') === 'asc') ? 'asc' : 'desc';
        $query->orderBy($sortBy, $sortDir);

        $perPage = min($request->input('per_page', 15), 100);
        $workOrders = $query->paginate($perPage);

        return response()->json($workOrders);
    }

    /**
     * Store a newly created work order.
     *
     * Creates a work order for an assembly product and auto-populates
     * items from the product's bill of materials.
     *
     * @param  Request  $request  The incoming HTTP request containing work order data
     */
    public function store(StoreWorkOrderRequest $request): JsonResponse
    {
        $organizationId = $request->user()->organization_id;

        $validated = $request->validated();

        // Verify the product is an assembly type
        $product = Product::where('id', $validated['product_id'])
            ->forOrganization($organizationId)
            ->firstOrFail();

        if ($product->type !== 'assembly') {
            return response()->json([
                'message' => 'Work orders can only be created for assembly products.',
                'error' => 'invalid_product_type',
            ], 422);
        }

        // Load the product's components
        $components = $product->components()->with('componentProduct')->get();

        if ($components->isEmpty()) {
            return response()->json([
                'message' => 'This assembly has no components defined. Add components before creating a work order.',
                'error' => 'no_components',
            ], 422);
        }

        // Retry on a work_order_number unique collision (concurrent creates in
        // the same org read the same MAX+1); nothing serialises them otherwise.
        $workOrder = SequenceNumberRetry::create(fn () => DB::transaction(function () use ($validated, $organizationId, $request, $components) {
            $workOrder = WorkOrder::create([
                'organization_id' => $organizationId,
                'product_id' => $validated['product_id'],
                'created_by' => $request->user()->id,
                'warehouse_id' => $validated['warehouse_id'] ?? null,
                'work_order_number' => WorkOrder::generateWorkOrderNumber($organizationId),
                'quantity' => $validated['quantity'],
                'status' => 'draft',
                'notes' => $validated['notes'] ?? null,
            ]);

            // Auto-populate items from the assembly's BOM * requested quantity
            foreach ($components as $component) {
                WorkOrderItem::create([
                    'work_order_id' => $workOrder->id,
                    'product_id' => $component->component_product_id,
                    'quantity_required' => $component->quantity * $validated['quantity'],
                    'quantity_consumed' => 0,
                ]);
            }

            return $workOrder;
        }));

        $workOrder->load(['product:id,name,sku,thumbnail', 'creator:id,name', 'items.product:id,name,sku,stock']);

        return response()->json([
            'message' => 'Work order created successfully',
            'data' => $workOrder,
        ], 201);
    }

    /**
     * Display the specified work order.
     *
     * @param  Request  $request  The incoming HTTP request
     * @param  WorkOrder  $workOrder  The work order to display
     */
    public function show(Request $request, WorkOrder $workOrder): JsonResponse
    {
        if ($workOrder->organization_id !== $request->user()->organization_id) {
            return response()->json([
                'message' => 'Work order not found',
                'error' => 'not_found',
            ], 404);
        }

        $workOrder->load([
            'product:id,name,sku,thumbnail,type',
            'creator:id,name',
            'items.product:id,name,sku,stock,thumbnail',
        ]);

        return response()->json([
            'data' => $workOrder,
        ]);
    }

    /**
     * Start a work order.
     *
     * Changes status from draft/pending to in_progress after verifying
     * all components have sufficient stock.
     *
     * @param  Request  $request  The incoming HTTP request
     * @param  WorkOrder  $workOrder  The work order to start
     */
    public function start(Request $request, WorkOrder $workOrder): JsonResponse
    {
        if ($workOrder->organization_id !== $request->user()->organization_id) {
            return response()->json([
                'message' => 'Work order not found',
                'error' => 'not_found',
            ], 404);
        }

        if (! in_array($workOrder->status, ['draft', 'pending'])) {
            return response()->json([
                'message' => 'Only draft or pending work orders can be started.',
                'error' => 'invalid_status',
            ], 422);
        }

        // Check component stock availability
        $workOrder->load('items.product');

        foreach ($workOrder->items as $item) {
            $remaining = $item->quantity_required - $item->quantity_consumed;
            if ($item->product->stock < $remaining) {
                return response()->json([
                    'message' => "Insufficient stock for component '{$item->product->name}'. Available: {$item->product->stock}, Required: {$remaining}.",
                    'error' => 'insufficient_stock',
                ], 422);
            }
        }

        $workOrder->update([
            'status' => 'in_progress',
            'started_at' => now(),
        ]);

        $workOrder->load(['product:id,name,sku,thumbnail', 'creator:id,name']);

        return response()->json([
            'message' => 'Work order started successfully',
            'data' => $workOrder,
        ]);
    }

    /**
     * Complete a work order.
     *
     * Decrements component stock and increments assembly product stock.
     *
     * @param  Request  $request  The incoming HTTP request
     * @param  WorkOrder  $workOrder  The work order to complete
     */
    public function complete(Request $request, WorkOrder $workOrder): JsonResponse
    {
        if ($workOrder->organization_id !== $request->user()->organization_id) {
            return response()->json([
                'message' => 'Work order not found',
                'error' => 'not_found',
            ], 404);
        }

        if ($workOrder->status !== 'in_progress') {
            return response()->json([
                'message' => 'Only in-progress work orders can be completed.',
                'error' => 'invalid_status',
            ], 422);
        }

        $validated = $request->validate([
            'quantity_produced' => ['nullable', 'integer', 'min:1', 'max:'.$workOrder->quantity],
        ]);

        $quantityProduced = $validated['quantity_produced'] ?? $workOrder->quantity;

        try {
            DB::transaction(function () use ($workOrder, $quantityProduced) {
                // Lock and re-read the row so a concurrent complete cannot pass
                // the pre-transaction guard and double-consume/double-produce
                // stock. Re-assert the status against the locked instance.
                $locked = WorkOrder::whereKey($workOrder->getKey())->lockForUpdate()->firstOrFail();

                if ($locked->status !== 'in_progress') {
                    throw new \RuntimeException('Only in-progress work orders can be completed.');
                }

                $locked->load('items.product');

                // Lock every product this completion touches (components +
                // assembly) up front, ordered by id, BEFORE any bin/stock
                // mutation, so bin operations run product-locked-first and this
                // path cannot ABBA-deadlock with a concurrent sale of a
                // component (order creation locks product then bins).
                $productIds = $locked->items->pluck('product_id')
                    ->push($locked->product_id)
                    ->filter()
                    ->unique()
                    ->sort()
                    ->values();
                Product::whereIn('id', $productIds)
                    ->where('organization_id', $locked->organization_id)
                    ->lockForUpdate()
                    ->get();

                // Decrement component stock, refusing to drive any component
                // negative — availability may have changed while the WO sat in
                // progress. A shortfall aborts the whole completion.
                foreach ($locked->items as $item) {
                    // Consume in proportion to what was actually produced, not
                    // the full planned quantity.
                    $targetConsumed = (int) ceil($item->quantity_required * ($quantityProduced / $locked->quantity));
                    $consumeQtyInt = $targetConsumed - (int) $item->quantity_consumed;

                    if ($consumeQtyInt > 0) {
                        // Draw the consumed components out of their location bins
                        // before the total drops (unbinned components seed from
                        // their pre-consumption stock).
                        app(ProductLocationStockService::class)->consume($item->product, $consumeQtyInt);

                        StockAdjustment::adjust(
                            $item->product,
                            -$consumeQtyInt,
                            'assembly_consumption',
                            "Consumed for work order {$locked->work_order_number}",
                            "Assembly of {$locked->product->name}",
                            $locked,
                            allowNegative: false
                        );

                        $item->update(['quantity_consumed' => $targetConsumed]);
                    }
                }

                // Increment assembly product stock
                StockAdjustment::adjust(
                    $locked->product,
                    $quantityProduced,
                    'assembly_production',
                    "Produced from work order {$locked->work_order_number}",
                    "Assembled {$quantityProduced} units",
                    $locked
                );

                // Book the assembled units into the product's primary bin.
                app(ProductLocationStockService::class)->receive($locked->product, $quantityProduced);

                $locked->update([
                    'status' => 'completed',
                    'quantity_produced' => $quantityProduced,
                    'completed_at' => now(),
                ]);
            });
        } catch (InsufficientStockException $e) {
            return response()->json([
                'message' => $e->getMessage(),
                'error' => 'insufficient_stock',
            ], 422);
        } catch (\RuntimeException $e) {
            return response()->json([
                'message' => $e->getMessage(),
                'error' => 'invalid_status',
            ], 422);
        }

        $workOrder->refresh()->load(['product:id,name,sku,thumbnail', 'creator:id,name']);

        return response()->json([
            'message' => "Work order completed. {$quantityProduced} units produced.",
            'data' => $workOrder,
        ]);
    }

    /**
     * Cancel a work order.
     *
     * If the work order was in progress, restores any consumed component stock.
     *
     * @param  Request  $request  The incoming HTTP request
     * @param  WorkOrder  $workOrder  The work order to cancel
     */
    public function cancel(Request $request, WorkOrder $workOrder): JsonResponse
    {
        if ($workOrder->organization_id !== $request->user()->organization_id) {
            return response()->json([
                'message' => 'Work order not found',
                'error' => 'not_found',
            ], 404);
        }

        if (! in_array($workOrder->status, ['draft', 'pending', 'in_progress'])) {
            return response()->json([
                'message' => 'Only draft, pending, or in-progress work orders can be cancelled.',
                'error' => 'invalid_status',
            ], 422);
        }

        try {
            DB::transaction(function () use ($workOrder) {
                // Lock and re-read so a cancel can't race a concurrent complete
                // (each would otherwise act on a stale status).
                $locked = WorkOrder::whereKey($workOrder->getKey())->lockForUpdate()->firstOrFail();

                if (! in_array($locked->status, ['draft', 'pending', 'in_progress'], true)) {
                    throw new \RuntimeException('Only draft, pending, or in-progress work orders can be cancelled.');
                }

                // If in_progress, restore any consumed component stock
                if ($locked->status === 'in_progress') {
                    $locked->load('items.product');

                    foreach ($locked->items as $item) {
                        $consumedQty = (int) ceil((float) $item->quantity_consumed);

                        if ($consumedQty > 0) {
                            StockAdjustment::adjust(
                                $item->product,
                                $consumedQty,
                                'assembly_reversal',
                                "Reversed for cancelled work order {$locked->work_order_number}",
                                'Work order cancelled - restoring consumed stock',
                                $locked
                            );

                            // Return the components to their primary bin.
                            app(ProductLocationStockService::class)->receive($item->product, $consumedQty);

                            $item->update(['quantity_consumed' => 0]);
                        }
                    }
                }

                $locked->update(['status' => 'cancelled']);
            });
        } catch (\RuntimeException $e) {
            return response()->json([
                'message' => $e->getMessage(),
                'error' => 'invalid_status',
            ], 422);
        }

        $workOrder->refresh()->load(['product:id,name,sku,thumbnail', 'creator:id,name']);

        return response()->json([
            'message' => 'Work order has been cancelled',
            'data' => $workOrder,
        ]);
    }

    /**
     * Delete a work order.
     *
     * Only draft or cancelled work orders, which have never moved stock, can
     * be deleted; anything else is a 422 `invalid_status`.
     */
    public function destroy(Request $request, WorkOrder $workOrder, WorkOrderService $workOrders): JsonResponse
    {
        if ($workOrder->organization_id !== $request->user()->organization_id) {
            return response()->json([
                'message' => 'Work order not found',
                'error' => 'not_found',
            ], 404);
        }

        try {
            $workOrders->delete($workOrder);
        } catch (InvalidStateException $e) {
            return response()->json(['message' => $e->getMessage(), 'error' => $e->errorCode], 422);
        }

        return response()->json(['message' => 'Work order deleted successfully']);
    }
}