# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Inventoros REST API
  version: 1.0.0
  description: |
    Inventoros provides a RESTful API for managing inventory, products, orders,
    suppliers, purchase orders, warehouses and related operations.

    ## Authentication

    All endpoints (except `POST /api/v1/login`) require a Bearer token.
    Authentication is implemented with Laravel Sanctum (personal access tokens).

    Obtain a token by calling `POST /api/v1/login` and then include it in the
    `Authorization` header:

    ```
    Authorization: Bearer {your-token}
    ```

    ## Multi-Tenancy

    Every resource is scoped to the authenticated user's organization via the
    `organization_id` foreign key. Cross-organization access returns `404`.

    ## Rate Limiting

    - All `/api/v1/*` routes: `60` requests per minute, keyed by user id (or
      IP for guests). Header: `throttle:api`.
    - `POST /api/v1/login`: `5` requests per minute, keyed by IP. Header:
      `throttle:5,1`.

    Exceeded throttling returns `429 Too Many Requests`.

    ## Pagination

    List endpoints (those returning `AnonymousResourceCollection`) are
    paginated and follow Laravel's standard JSON envelope:

    ```json
    {
      "data": [ ... ],
      "links": { "first": "...", "last": "...", "prev": null, "next": "..." },
      "meta": { "current_page": 1, "last_page": 3, "per_page": 15, "total": 42 }
    }
    ```

    Use `?per_page=` (max `100`) and `?page=` query params.

    ## Permissions

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /orders/{order}:
    parameters:
      - { name: order, in: path, required: true, schema: { type: integer } }
    get:
      tags: [Orders]
      summary: Show an order
      responses:
        '200': { description: Order, content: { application/json: { schema: { type: object, properties: { data: { $ref: '#/components/schemas/Order' } } } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Orders]
      summary: Update an order's metadata or status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_name: { type: string, maxLength: 255 }
                customer_email: { type: string, format: email, maxLength: 255 }
                customer_address: { type: string }
                status: { type: string, enum: [pending, processing, shipped, delivered, cancelled] }
                notes: { type: string }
                metadata: { type: object, additionalProperties: true }
      responses:
        '200': { description: Updated }
    delete:
      tags: [Orders]
      summary: Delete an order (restores product stock)
      responses:
        '200': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /work-orders/{workOrder}:
    parameters:
      - { name: workOrder, in: path, required: true, schema: { type: integer } }
    get:
      tags: [Work Orders]
      summary: Show a work order
      responses: { '200': { description: Work order }, '404': { $ref: '#/components/responses/NotFound' } }
    delete:
      tags: [Work Orders]
      summary: Delete a draft or cancelled work order
      description: "Requires `manage_stock`. Work orders that have started or completed cannot be deleted."
      responses:
        '200': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { description: "Not draft or cancelled (`invalid_status`)" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: PlainText
      description: |
        Laravel Sanctum personal access token, in the format
        `<id>|<random>` returned by `POST /api/v1/login`.
...
  responses:
...
    NotFound:
      description: Resource not found in the user's organization
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              error: { type: string, example: not_found }
...
  schemas:
...
    Order:
      type: object
      properties:
        id: { type: integer }
        order_number: { type: string }
        source: { type: string, nullable: true }
        external_id: { type: string, nullable: true }
        customer_name: { type: string }
        customer_email: { type: string, format: email, nullable: true }
        customer_address: { type: string, nullable: true }
        status: { type: string, enum: [pending, processing, shipped, delivered, cancelled] }
        subtotal: { type: number }
        tax: { type: number }
        shipping: { type: number }
        total: { type: number }
        currency: { type: string }
        order_date: { type: string, format: date-time, nullable: true }
        shipped_at: { type: string, format: date-time, nullable: true }
        delivered_at: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
        metadata: { type: object, additionalProperties: true, nullable: true }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
        items_count: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrderItem:
      type: object
      properties:
        id: { type: integer }
        order_id: { type: integer }
        product_id: { type: integer }
        product_name: { type: string }
        sku: { type: string, nullable: true }
        quantity: { type: integer }
        unit_price: { type: number }
        subtotal: { type: number }
        tax: { type: number }
        total: { type: number }
        metadata: { type: object, additionalProperties: true, nullable: true }
...